# apps/next-frontend/public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ticketeate API",
    "version": "1.0.0",
    "description": "Documentación de la API de Ticketeate (OAS 3.0)."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Desarrollo local" }
  ],
...
  "paths": {
...
    "/api/tickets/send-email": {
      "post": {
        "tags": ["Tickets"],
        "summary": "Enviar ticket por email",
        "security": [{ "bearerAuth": [] }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Enviado" } }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  }
}